# Eliminated bounds checks inside parsing a field.