Each field parser gets 16 bytes of slop.  This requires using a patch
buffer at end-of-buffer.

This addes 80% of what is needed to support a pull parser with a data
callback, since the main parser is now tolerant to buffer flips.

Combined with an optimization to long varint parsing, the performance
is a net improvement, though we are taking a noticeable 25% hit to code
side in the decoder.

```
name                                 old time/op  new time/op  delta
ArenaOneAlloc                        21.1ns ± 0%  21.3ns ± 0%  +1.33%  (p=0.000 n=12+12)
ArenaInitialBlockOneAlloc            6.02ns ± 0%  6.02ns ± 0%    ~     (p=0.579 n=10+10)
LoadDescriptor_Upb                    111µs ± 1%   110µs ± 1%  -0.91%  (p=0.003 n=11+12)
LoadDescriptor_Proto2                 258µs ± 1%   258µs ± 1%    ~     (p=0.674 n=10+12)
Parse_Upb_FileDesc_WithArena         11.2µs ± 0%  10.4µs ± 0%  -6.67%  (p=0.000 n=12+12)
Parse_Upb_FileDesc_WithInitialBlock  10.6µs ± 0%  10.1µs ± 0%  -4.48%  (p=0.000 n=12+11)
SerializeDescriptor_Proto2           5.36µs ± 5%  5.36µs ± 3%    ~     (p=0.880 n=12+11)
SerializeDescriptor_Upb              11.9µs ± 0%  12.0µs ± 0%  +0.81%  (p=0.000 n=12+12)

    FILE SIZE        VM SIZE
 --------------  --------------
   +23% +1.11Ki   +24% +1.06Ki    upb/decode.c
     +15%    +560   +15%    +560    decode_msg
    +140%    +240  +188%    +240    decode_longvarint64
    [NEW]    +174  [NEW]    +128    decode_isdonefallback
     +56%    +160   +65%    +160    upb_decode
 -49.7% -1.06Ki  [ = ]       0    [Unmapped]
  +0.0%     +48  +0.9% +1.06Ki    TOTAL
```